backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["nwilson314 <nsw.wilson@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = {extras = ["standard"], version = "^0.115.7"}
sqlmodel = "^0.0.22"
alembic = "^1.14.1"
psycopg2-binary = "^2.9.10"
pydantic-settings = "^2.7.1"
loguru = "^0.7.3"
black = "^25.1.0"
python-dotenv = "^1.0.1"
pyjwt = "^2.10.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pytest = "^8.3.4"
openai = "^1.66.2"
yt-dlp = "^2025.2.19"
markdown = "^3.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
